pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "fitz>=0.0.1.dev2",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "pymupdf>=1.25.5",
    "requests>=2.32.3",
]
